added tests for checking file types and refactored to make it easier to maintain.
